Move all error building code into a dedicated class
