+++ src/BugsnagUnity/INativeClient.cs
@@ -1,4 +1,5 @@
 ﻿using BugsnagUnity.Payload;
+using System;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 
@@ -101,6 +102,14 @@ namespace BugsnagUnity
         void AddNativeMetadata(string section, IDictionary<string, object> data);
 
         IDictionary<string, object> GetNativeMetadata();
+
+        /// <summary>
+        /// Find the native loaded image that corresponds to a native instruction address
+        /// supplied by il2cpp_native_stack_trace().
+        /// </summary>
+        /// <param name="address">The address to find the corresponding image of</param>
+        /// <returns>The corresponding image, or null</returns>
+        LoadedImage FindImageAtAddress(UInt64 address);
 
         bool ShouldAttemptDelivery();
