+++ src/BugsnagUnity/Native/Android/NativeClient.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using BugsnagUnity.Payload;
 using UnityEngine;
@@ -160,6 +161,11 @@ namespace BugsnagUnity
         public void RegisterForOnSessionCallbacks()
         {
             NativeInterface.RegisterForOnSessionCallbacks();
+        }
+
+        public LoadedImage FindImageAtAddress(UInt64 address)
+        {
+            return null;
         }
     }
