+++ src/BugsnagUnity/Native/Fallback/NativeClient.cs
@@ -169,5 +169,10 @@ namespace BugsnagUnity
         {
             // Not Used on this platform
         }
+
+        public LoadedImage FindImageAtAddress(UInt64 address)
+        {
+            return null;
+        }
     }
 }
