+++ src/BugsnagUnity/Native/Windows/NativeClient.cs
@@ -216,5 +216,10 @@ namespace BugsnagUnity
         {
             // Not Used on this platform
         }
+
+        public LoadedImage FindImageAtAddress(UInt64 address)
+        {
+            return null;
+        }
     }
 }
